requirements.txt
dash
pandas
plotly
requests
gunicorn
torch
Pillow
scikit-learn
-e ./morphomics_exp